#  Adapt pixel cpe algo to better handle broken clusters [15.0.X]