--- a/CondFormats/SiPixelTransient/src/SiPixelUtils.cc
+++ b/CondFormats/SiPixelTransient/src/SiPixelUtils.cc
@@ -13,21 +13,23 @@ namespace siPixelUtils {
   //!  X and Y, in the interest of the simplicity of the code, all parameters
   //!  are passed by the caller.
   //-----------------------------------------------------------------------------
-  float generic_position_formula(int size,                    //!< Size of this projection.
-                                 int q_f,                     //!< Charge in the first pixel.
-                                 int q_l,                     //!< Charge in the last pixel.
-                                 float upper_edge_first_pix,  //!< As the name says.
-                                 float lower_edge_last_pix,   //!< As the name says.
-                                 float lorentz_shift,         //!< L-shift at half thickness
-                                 float theThickness,          //detector thickness
-                                 float cot_angle,             //!< cot of alpha_ or beta_
-                                 float pitch,                 //!< thePitchX or thePitchY
-                                 float pitchfraction_first,
-                                 float pitchfraction_last,
-                                 float eff_charge_cut_low,   //!< Use edge if > w_eff  &&&
-                                 float eff_charge_cut_high,  //!< Use edge if < w_eff  &&&
-                                 float size_cut              //!< Use edge when size == cuts
-  ) {
+  float generic_position_formula(
+      int size,                    //!< Size of this projection.
+      int q_f,                     //!< Charge in the first pixel.
+      int q_l,                     //!< Charge in the last pixel.
+      float upper_edge_first_pix,  //!< As the name says.
+      float lower_edge_last_pix,   //!< As the name says.
+      float lorentz_shift,         //!< L-shift at half thickness
+      float theThickness,          //detector thickness
+      float cot_angle,             //!< cot of alpha_ or beta_
+      float pitch,                 //!< thePitchX or thePitchY
+      float pitchfraction_first,
+      float pitchfraction_last,
+      float eff_charge_cut_low,   //!< Use edge if > w_eff  &&&
+      float eff_charge_cut_high,  //!< Use edge if < w_eff  &&&
+      float size_cut,             //!< Use edge when size == cuts
+      float delta_length_cut,     //!< if charge len - cls size > this (in pix), use one-sided reco
+      bool goodEdgeAlgo) {
     float geom_center = 0.5f * (upper_edge_first_pix + lower_edge_last_pix);
 
     //--- The case of only one pixel in this projection is separate.  Note that
@@ -56,11 +58,26 @@ namespace siPixelUtils {
     //--- it with an *average* effective charge width, which is the average
     //--- length of the edge pixels.
     //
-    //  bool usedEdgeAlgo = false;
     if ((size >= size_cut) || ((w_eff / pitch < eff_charge_cut_low) | (w_eff / pitch > eff_charge_cut_high))) {
       w_eff = pitch * 0.5f * sum_of_edge;  // ave. length of edge pixels (first+last) (cm)
-                                           //  usedEdgeAlgo = true;
-    }
+
+      if (goodEdgeAlgo) {
+        float delta = w_eff - 0.5 * sum_of_edge * pitch;
+        if (delta / pitch > delta_length_cut) {
+          // define the centers of the first last last pixel coordinates
+          float x1 = upper_edge_first_pix - 0.5 * pitchfraction_first * pitch;
+          float x2 = lower_edge_last_pix + 0.5 * pitchfraction_last * pitch;
+          //  observed cluster is much shorter than expected, use one-sided reco
+          if (w_pred > 0.f) {
+            float hit_pos = x1 + 0.5 * w_pred;
+            return hit_pos;
+          } else {
+            float hit_pos = x2 + 0.5 * w_pred;
+            return hit_pos;
+          }
+        }  //if (delta / pitch > delta_length_cut)
+      }  //if(goodEdgeAlgo)
+    }  //if(size >= size_cut) ||...
 
     //--- Finally, compute the position in this projection
     float q_diff = q_l - q_f;