--- a/RecoLocalTracker/SiPixelRecHits/python/PixelCPEClusterRepair_cfi.py
+++ b/RecoLocalTracker/SiPixelRecHits/python/PixelCPEClusterRepair_cfi.py
@@ -6,3 +6,8 @@ templates2_speed0 = _templates2_default.clone(
     ComponentName = "PixelCPEClusterRepairWithoutProbQ",
     speed = 0
 )
+
+# Enable the good edge algorithm in pixel hit reconstruction that handles broken/truncated pixel cluster caused by radiation damage
+from Configuration.ProcessModifiers.siPixelGoodEdgeAlgo_cff import siPixelGoodEdgeAlgo
+siPixelGoodEdgeAlgo.toModify(templates2, GoodEdgeAlgo = True)
+siPixelGoodEdgeAlgo.toModify(templates2_speed0, GoodEdgeAlgo = True)