… RecoLocalTracker/SiPixelRecHits/src/PixelCPEClusterRepair.cc
… RecoLocalTracker/SiPixelRecHits/src/PixelCPEFastParamsHost.cc
… RecoLocalTracker/SiPixelRecHits/src/PixelCPEGeneric.cc
… RecoLocalTracker/SiPixelRecHits/src/PixelCPETemplateReco.cc
… RecoLocalTracker/SiPixelRecHits/src/SiPixelTemplateReco.cc