@mmusich 2025-05-05
assign heterogeneous 
   * for the pixel local reco on GPU part