@mmusich 2025-05-05
backport of https://github.com/cms-sw/cmssw/pull/47966