@mmusich 2025-05-05
enable gpu