@mroguljic 2025-05-06
`runTheMatrix` finished with `50 49 46 41 19 1 1 1 1 1 1 tests passed, 0 0 1 0 0 0 0 0 0 0 0 failed`. Workflow 1000.0 fails at step3 with
```
----- Begin Fatal Exception 06-May-2025 07:46:12 CEST-----------------------
An exception of category 'FatalRootError' occurred while
   [0] Calling EventProcessor::runToCompletion (which does almost everything after beginJob and before endJob)
   Additional Info:
      [a] Fatal Root Error: @SUB=TStorageFactorySystem::Unlink
Unsupported

----- End Fatal Exception -------------------------------------------------
```
Not sure if this is related to the changes.