@mroguljic 2025-05-06
> > Not sure if this is related to the changes.
> 
> @mroguljic, do you happen to have run your tests on `eos`? This looks related to #44369, thus unrelated.

Thank you Marco. Indeed, I ran this on `eos`