@vlimant 2025-05-08
maybe synchronise the PR, while timing testing is on-going