@fwyzard 2025-05-08
+heterogeneous