@elusian 2025-05-08
Hello, I'm running an HLT timing measurement using
- /dev/CMSSW_15_0_0/GRun/V76
- 2024 data
- storm customizer for the L1 menu
- GT 150X_dataRun3_HLT_Candidate_2025_05_08_08_50_15
- configuration change with
```
process.hltESPPixelCPEGeneric.GoodEdgeAlgo = cms.bool(True)
process.hltESPPixelCPEFastParamsPhase1.GoodEdgeAlgo = cms.bool(True)
```
I will write here as soon as the studies are complete