@elusian 2025-05-08
These are the results:
- [Baseline](https://cmshlttiming.app.cern.ch/display/elusiani/CMSSW_15_0_5_Tk25.20250508_151410): 479.1 ms
- [New CPE algo](https://cmshlttiming.app.cern.ch/display/elusiani/CMSSW_15_0_5_Tk25_newPixelCPE.20250508_151559): 489.3 ms

To summarize: 
- Pixel GPU and pixel conversion do not change much
- Pixel, mostly the HitPair producer, increases by 1 ms
- Tracking increases by 10 ms, mostly in the Ckf  track candidate maker (e.g. the doublet recovery + the POG specific sequences like what is used by tau or for displaced tracks)