@cmsbuild 2025-05-08
+1

**Size**: This PR adds an extra 28KB to repository
**Summary**: https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-7b81a1/45961/summary.html
**COMMIT**: 85e787d8af60f8c60de77f546c0b1447d96e34db
**CMSSW**: CMSSW_15_0_X_2025-05-08-1100/el8_amd64_gcc12
**Additional Tests**: CUDA,ROCM
**User test area**: For local testing, you can use `/cvmfs/cms-ci.cern.ch/week0/cms-sw/cmssw/48008/45961/install.sh` to create a dev area with all the needed externals and cmssw changes.


## Comparison Summary



**Summary**:
 - You potentially [removed 22 lines](https://cmssdt.cern.ch/SDT/jenkins-artifacts/baseLineComparisons/CMSSW_15_0_X_2025-05-08-1100+7b81a1/68353/validateJR/logRootQA-events.log) from the logs
 - [Reco comparison results: 3676 differences found in the comparisons](https://cmssdt.cern.ch/SDT/jenkins-artifacts/baseLineComparisons/CMSSW_15_0_X_2025-05-08-1100+7b81a1/68353/validateJR.html)
 - DQMHistoTests: Total files compared: 50
 - DQMHistoTests: Total histograms compared: 4012378
 - DQMHistoTests: [Total failures: 17544](https://cmssdt.cern.ch/SDT/jenkins-artifacts/baseLineComparisons/CMSSW_15_0_X_2025-05-08-1100+7b81a1/68353/validateJR.html)
 - DQMHistoTests: Total nulls: 0
 - DQMHistoTests: Total successes: 3994814
 - DQMHistoTests: Total skipped: 20
 - DQMHistoTests: Total Missing objects: 0
 - DQMHistoSizes: Histogram memory added: 0.0 KiB( 49 files compared)
 - Checked 218 log files, 189 edm output root files, 50 DQM output files
 - [TriggerResults](https://cmssdt.cern.ch/SDT/jenkins-artifacts/baseLineComparisons/CMSSW_15_0_X_2025-05-08-1100+7b81a1/68353/triggerResults): no differences found


## CUDA Comparison Summary



**Summary**:
 - No significant changes to the logs found
 - [Reco comparison results: 8 differences found in the comparisons](https://cmssdt.cern.ch/SDT/jenkins-artifacts/baseLineComparisonsCUDA/CMSSW_15_0_X_2025-05-08-1100+7b81a1/68352/validateJR.html)
 - DQMHistoTests: Total files compared: 7
 - DQMHistoTests: Total histograms compared: 53071
 - DQMHistoTests: [Total failures: 379](https://cmssdt.cern.ch/SDT/jenkins-artifacts/baseLineComparisonsCUDA/CMSSW_15_0_X_2025-05-08-1100+7b81a1/68352/validateJR.html)
 - DQMHistoTests: Total nulls: 0
 - DQMHistoTests: Total successes: 52692
 - DQMHistoTests: Total skipped: 0
 - DQMHistoTests: Total Missing objects: 0
 - DQMHistoSizes: Histogram memory added: 0.0 KiB( 6 files compared)
 - Checked 24 log files, 30 edm output root files, 7 DQM output files
 - [TriggerResults](https://cmssdt.cern.ch/SDT/jenkins-artifacts/baseLineComparisonsCUDA/CMSSW_15_0_X_2025-05-08-1100+7b81a1/68352/triggerResults): no differences found


## ROCM Comparison Summary



**Summary**:
 - No significant changes to the logs found
 - [Reco comparison results: 0 differences found in the comparisons](https://cmssdt.cern.ch/SDT/jenkins-artifacts/baseLineComparisonsROCM/CMSSW_15_0_X_2025-05-08-1100+7b81a1/68355/validateJR.html)
 - DQMHistoTests: Total files compared: 7
 - DQMHistoTests: Total histograms compared: 53071
 - DQMHistoTests: [Total failures: 32](https://cmssdt.cern.ch/SDT/jenkins-artifacts/baseLineComparisonsROCM/CMSSW_15_0_X_2025-05-08-1100+7b81a1/68355/validateJR.html)
 - DQMHistoTests: Total nulls: 0
 - DQMHistoTests: Total successes: 53039
 - DQMHistoTests: Total skipped: 0
 - DQMHistoTests: Total Missing objects: 0
 - DQMHistoSizes: Histogram memory added: 0.0 KiB( 6 files compared)
 - Checked 24 log files, 30 edm output root files, 7 DQM output files