@perrotta 2025-05-09
assign hlt
(so that TSG can sign off as soon as they are satisfied with timing measurements)