@mmusich 2025-05-09
>  (so that TSG can sign off as soon as they are satisfied with timing measurements)

the timing measurements reported here https://github.com/cms-sw/cmssw/pull/48008#issuecomment-2863390486 indicate a ~ 2% increase of the timing, which is not negligible. It would be nice to have physics validation as well to be able to gauge the impact in terms of physics before taking a decision. 
OTOH, the PR per se shouldn't change the status quo, as the modification is not active by default, so technically we could go ahead as it is now and have the validation later before actually changing the menu (via configuration change).
@missirol @mtosi please chime in if necessary.