@missirol 2025-05-09
I agree with what Marco wrote in https://github.com/cms-sw/cmssw/pull/48008#issuecomment-2866844258.

We'll follow up with STEAM, Tracker DPG and Tracking POG for the validation of the HLT performance.

(I guess we do not expect any impact on HLT throughput if the new option is not enabled, but if you think this is worth checking explicitly, please let me know)