@fwyzard 2025-05-09
The change to the CPE GPU code is small, but not negligible. To be on the safe side, I would check the impact on the HLT even with the new setting disabled in the configuration.